revert android stuff
