+++ android/src/main/java/com/RNJwplayer/RNJwplayerModule.java
@@ -1,0 +1,25 @@
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContextBaseJavaModule;
+import com.facebook.react.bridge.ReactMethod;
+import com.facebook.react.bridge.Callback;
+
+public class RNJwplayerModule extends ReactContextBaseJavaModule {
+
+    private final ReactApplicationContext reactContext;
+
+    public RNJwplayerModule(ReactApplicationContext reactContext) {
+        super(reactContext);
+        this.reactContext = reactContext;
+    }
+
+  @Override
+  public String getName() {
+    return "RNJwplayer";
+  }
+
+    @ReactMethod
+    public void sampleMethod(String stringArgument, int numberArgument, Callback callback) {
+        // TODO: Implement some real useful functionality
+        callback.invoke("Received numberArgument: " + numberArgument + " stringArgument: " + stringArgument);
+    }
+}
